XML topology file parsing and TEDB operations
git-svn-id: file:///var/www/svn/mxtce-sw@47 52860be4-34ab-47e5-b986-9ec2afffd7f9

+++ src/tedbManager/resource.cc
@@ -82,7 +82,7 @@ void Port::AddLink(Link* link)
     this->links[link->GetName()] = link;
 }
 
-list<Port*> NodeIfAdaptMatrix::GetAdaptToPorts(Port* fromPort, float bw)
+list<Port*> NodeIfAdaptMatrix::GetAdaptToPorts(Port* fromPort, long bw)
 {
     list<Port*> portList;
     if (portMap.find(fromPort) != portMap.end())
@@ -101,7 +101,7 @@ list<Port*> NodeIfAdaptMatrix::GetAdaptToPorts(Port* fromPort, float bw)
 }
 
 
-list<Port*> NodeIfAdaptMatrix::GetAdaptFromPorts(Port* toPort, float bw)
+list<Port*> NodeIfAdaptMatrix::GetAdaptFromPorts(Port* toPort, long bw)
 {
     list<Port*> portList;
     if (portMap.find(toPort) != portMap.end())
